backend/requirements.txt
pandas
numpy
geopandas
supabase
python-dotenv
